+++ src/components/goals/GoalTitle.js
@@ -3,7 +3,7 @@ import PropTypes from 'prop-types';
 import Icon from 'components/common/Icon';
 import { useGoal } from 'hooks/UseGoal';
 
-export function GoalTitle(props) {
+function GoalTitle(props) {
     const goal = useGoal(props.goalId);
     return goal ? <Icon icon="circle" color={goal.color} text={goal.title} /> : <span>&nbsp;</span>;
 }