+++ src/components/notefields/NoteFieldTitle.js
@@ -3,7 +3,7 @@ import PropTypes from 'prop-types';
 import Icon from 'components/common/Icon';
 import { useNoteField } from 'hooks/UseNoteField';
 
-export function NoteFieldTitle(props) {
+function NoteFieldTitle(props) {
     const noteField = useNoteField(props.noteFieldId);
     return noteField ? <Icon icon="circle" color={noteField.color} text={noteField.title} /> : <span>&nbsp;</span>;
 }